Simplify typecasts using unsigned constant suffixes and better types
Use U-suffixed constants (1U, 255U, 12U, 20U) instead of explicit
(uint32_t) casts on expressions. Change CMYK local variables from
uint16_t/uint32_t to unsigned int to match the natural arithmetic
type. Remove redundant intermediate casts: inner (int16_t) before
shift in tif_luv.c, unnecessary (size_t) wrapping in tif_dirread.c,
and redundant (double) on already-double variables in tiff2ps.c.

+++ libtiff/tif_dirread.c
@@ -5315,7 +5315,7 @@ static void TIFFReadDirectoryCheckOrder(TIFF *tif, TIFFDirEntry *dir,
                             "ascending order");
             break;
         }
-        m = (uint32_t)(o->tdir_tag + 1);
+        m = o->tdir_tag + 1U;
     }
 }
 
@@ -7835,7 +7835,7 @@ static int TIFFFetchStripThing(TIFF *tif, TIFFDirEntry *dir, uint32_t nstrips,
         }
         if (dir->tdir_count)
             _TIFFmemcpy(resizeddata, data,
-                        (tmsize_t)((size_t)((uint32_t)dir->tdir_count) * sizeof(uint64_t)));
+                        (tmsize_t)((uint32_t)dir->tdir_count * sizeof(uint64_t)));
         _TIFFmemset(resizeddata + (uint32_t)dir->tdir_count, 0,
                     (tmsize_t)((size_t)(nstrips - (uint32_t)dir->tdir_count) * sizeof(uint64_t)));
         _TIFFfreeExt(tif, data);
